Join all apis into one and split into logical parts
