+++ src/pages/Project/Overview.jsx
@@ -11,20 +11,22 @@ import Tasklist from '../../components/Tasklist';
 export default function Project() {
   const tasks = useGetTasksQuery();
   const projects = useGetProjectsQuery();
+  const heatmaps = useGetHeatmapsQuery();
   const { projectID } = useParams();
   const project =
     useGetProjectsQuery().data.find((projecto) => projecto._id == projectID) ??
     useGetProjectsQuery().data.find((projecto) => projecto.name == 'Default');
   const heatmap =
-    useGetHeatmapsQuery().data.find((heatmapo) => heatmapo.project._id == projectID) ??
-    [];
+    useGetHeatmapsQuery().data.find((heatmapo) => heatmapo.project._id == projectID)
+
+  console.log(heatmaps);
 
   const dateEnd = startOfDay(new Date());
   const dateStart = subYears(dateEnd, 1);
   const dayLength = 2;
   const yearData = YearDataSimple(dateStart);
 
-  return tasks.isFetching || projects.isFetching ? (
+  return tasks.isFetching || projects.isFetching || heatmaps.isFetching ? (
     <> </>
   ) : (
     <div className="project-overview" style={{ '--cell-height': '11px', '--cell-width': '11px' }}>
