+++ src/state/store.js
@@ -1,8 +1,6 @@
 import { configureStore } from '@reduxjs/toolkit';
 import { setupListeners } from '@reduxjs/toolkit/query';
-import { todolistApi } from './services/todolist';
-import { projectApi } from './services/project';
-import { heatmapApi } from './services/heatmap';
+import api from './services/api';
 import projectOverlayReducer from './features/projectOverlay/projectOverlaySlice';
 import taskOverlayReducer from './features/taskOverlay/taskOverlaySlice';
 import cellAddReducer from './features/cellAdd/cellAddSlice';
@@ -11,9 +9,7 @@ import themeReducer from './features/theme/themeSlice';
 const store = configureStore({
   reducer: {
     // Add the generated reducer as a specific top-level slice
-    [todolistApi.reducerPath]: todolistApi.reducer,
-    [projectApi.reducerPath]: projectApi.reducer,
-    [heatmapApi.reducerPath]: heatmapApi.reducer,
+    [api.reducerPath]: api.reducer,
     projectOverlay: projectOverlayReducer,
     taskOverlay: taskOverlayReducer,
     cellAdd: cellAddReducer,
@@ -23,9 +19,7 @@ const store = configureStore({
   // and other useful features of `rtk-query`.
   middleware: (getDefaultMiddleware) =>
     getDefaultMiddleware().concat([
-      todolistApi.middleware,
-      projectApi.middleware,
-      heatmapApi.middleware,
+      api.middleware,
     ]),
 });
